requirements.txt
Django>=4.2
djangorestframework>=3.14
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
Pillow>=10.0
django-cors-headers>=4.3
python-decouple>=3.8
qrcode>=7.4
opencv-python>=4.8
numpy>=1.24
pandas>=2.0
django-extensions>=3.2
celery>=5.3
redis>=4.6
channels>=4.0
channels-redis>=4.1
whitenoise>=6.5django-filter>=23.0
